Pretty major rework of how keybindings are triggered
Because Rainbow CSV didn't have a single context I could check on the
keybindings, I needed to rework how to determine if we're in a valid
tabnav context. Ultimately I settled on a custom view event listener
that listens for the 'is_tabnav_context' keybinding context. Along the
way I also added the ability to explicitly enable and disable tabnav on
a given view, and implemented the 'auto_csv' context to automatically
determine the appropriate CSV delimiter to use.

This has the added bonus of not needing duplicate keybindings for each
syntax. A single set of keybinding definitions works globally.