Standardize on left/right directions
Initialily was trying to be left/right agnostic, so that in principle
RTL languages could be supported, but it turns out that Sublime doesn't
really support RTL languages. So most likely if they did add RTL
language support, things would be broken in one way or another.

So, in light of that, "left" and "right" is clearer language.