Add support for OrgMode tables
Split the markup line pattern context setting into an array instead of a
single value, since Org Mode tables have three distinct types of markup
lines. In principle, this _could_ be handled by a single regex, but this
is cleaner, I think. Also, I might refactor the markup patterns further
to link a line pattern with a particular set of markup cell patterns.
For OrgMode, we're able to distinguish the three by the first cell, so
it's not strictly necessary, but I think it'll make it cleaner. (Also,
less regex parsing to do.)